replace RTMP/Twitch toggle with pronouns field
RTMP is no longer used, and the space is a good size for a pronouns text field, so let's just take care of both things at once. RTMP is still referenced by the replay buttons; I don't actually know how those work(ed).

Pronouns are written to a file (like streamers) so they can be automatically picked up as well, but are currently _not_ written to autocomplete.

Pronouns can be read from a macsg: protocol string, but are not required to be specified for each racer, e.g.:
  macsg:twitch,racer1|pronouns1,racer2,racer3|pronouns3,racer4